backend/vercel-requirements.txt
# Vercel-optimized requirements
fastapi
uvicorn[standard]
tensorflow-cpu
pillow
numpy
opencv-python-headless
python-multipart
